% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryatlas-package.R
\docType{package}
\name{countryatlas-package}
\alias{countryatlas}
\alias{countryatlas-package}
\title{countryatlas: join World Bank data, country codes and maps on the ISO spine}
\description{
\code{countryatlas} exists to kill one recurring source of pain: country names
never line up across data sources. The package makes ISO codes the universal
join key and hands you a ready-to-map tibble that stitches together map
geometry (\code{\link[ggplot2:map_data]{ggplot2::map_data()}} or Natural Earth \code{sf}), World Bank indicators
(\code{\link[WDI:WDI]{WDI::WDI()}}) and the \code{\link[countrycode:countrycode]{countrycode::countrycode()}} crosswalk.
}
\details{
The happy path stays one call: \code{\link[=world_data]{world_data()}}. Everything else is opt-in.
}
\section{Core data assembly}{

\code{\link[=world_data]{world_data()}}, \code{\link[=country_data]{country_data()}}, \code{\link[=world_geometry]{world_geometry()}}, \code{\link[=locate_country]{locate_country()}}.
}

\section{The join engine}{

\code{\link[=standardize_country]{standardize_country()}}, \code{\link[=join_world]{join_world()}}, \code{\link[=attach_geometry]{attach_geometry()}}, \code{\link[=country_join]{country_join()}},
\code{\link[=country_join_all]{country_join_all()}}.
}

\section{Diagnostics}{

\code{\link[=check_country_match]{check_country_match()}}, \code{\link[=repair_country_names]{repair_country_names()}}, \code{\link[=country_overrides]{country_overrides()}},
\code{\link[=audit_coverage]{audit_coverage()}}.
}

\section{Reference data}{

\code{\link[=convert_country]{convert_country()}}, \code{\link[=country_codes]{country_codes()}}, \code{\link[=country_groups]{country_groups()}}, \code{\link[=in_group]{in_group()}},
\code{\link[=wdi_search]{wdi_search()}}, and the datasets \link{country_meta}, \link{common_indicators},
\link{country_groups_tbl}, \link{world_snapshot}, \link{world_tiles}.
}

\section{Analysis helpers}{

\code{\link[=per_capita]{per_capita()}}, \code{\link[=aggregate_regions]{aggregate_regions()}}, \code{\link[=rank_countries]{rank_countries()}}, \code{\link[=complete_years]{complete_years()}},
\code{\link[=growth_rate]{growth_rate()}}, \code{\link[=index_to]{index_to()}}, \code{\link[=share_of_world]{share_of_world()}}.
}

\section{Visualization}{

\code{\link[=world_map]{world_map()}}, \code{\link[=globe_map]{globe_map()}}, \code{\link[=facet_map]{facet_map()}}, \code{\link[=bubble_map]{bubble_map()}}, \code{\link[=bivariate_map]{bivariate_map()}},
\code{\link[=cartogram_map]{cartogram_map()}}, \code{\link[=tile_map]{tile_map()}}, \code{\link[=flow_map]{flow_map()}}, \code{\link[=animate_world]{animate_world()}},
\code{\link[=interactive_map]{interactive_map()}}, \code{\link[=geom_country_labels]{geom_country_labels()}}, \code{\link[=theme_world_map]{theme_world_map()}}.
}

\section{Database rendering (ggsql)}{

\code{\link[=as_ggsql_source]{as_ggsql_source()}}, \code{\link[=world_query]{world_query()}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://pursuitofdatascience.github.io/countryatlas/}
  \item \url{https://github.com/PursuitOfDataScience/countryatlas}
  \item Report bugs at \url{https://github.com/PursuitOfDataScience/countryatlas/issues}
}

}
\author{
\strong{Maintainer}: Youzhi Yu \email{yuyouzhi666@icloud.com}

Authors:
\itemize{
  \item Youzhi Yu \email{yuyouzhi666@icloud.com}
}

}
\keyword{internal}
