% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{in_group}
\alias{in_group}
\title{Is a country in a group?}
\usage{
in_group(x, group, origin = "country.name")
}
\arguments{
\item{x}{A vector of country names or codes.}

\item{group}{A single group name (see \code{\link[=country_groups]{country_groups()}}).}

\item{origin}{How to read \code{x} (default \code{"country.name"}).}
}
\value{
A logical vector the same length as \code{x}.
}
\description{
A vectorised membership predicate built on \code{\link[=country_groups]{country_groups()}}.
}
\examples{
in_group(c("France", "United States", "Japan"), "EU")
}
