% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_data.R
\name{world_data}
\alias{world_data}
\title{Obtaining countries' information}
\usage{
world_data(year)
}
\arguments{
\item{year}{Year. The minimum input year is 1960.}
}
\value{
A tibble with country information around the world.
}
\description{
This function returns a tibble comprising geo coordinates of each country
with the iso2c and iso3c code, income and GDP per capita information, along
with which continent each country is classified with.
}
\examples{
\dontrun{
world_data(2020)
}
}
